all related to extension strategy renamed to reference it
