+++ internal/code-generator/assets/extensions.go.tmpl
