+++ internal/code-generator/generator/test_files/extensions.test.go.tmpl
