+++ internal/code-generator/main.go
@@ -8,11 +8,13 @@ import (
 )
 
 const (
-	// languages.go generation
-	languagesYAML     = ".linguist/lib/linguist/languages.yml"
-	langFile          = "languages.go"
-	languagesTmplPath = "internal/code-generator/assets/languages.go.tmpl"
-	languagesTmpl     = "languages.go.tmpl"
+	// languages info file
+	languagesYAML = ".linguist/lib/linguist/languages.yml"
+
+	// extensions_map.go generation
+	extensionsFile     = "extensions_map.go"
+	extensionsTmplPath = "internal/code-generator/assets/extensions.go.tmpl"
+	extensionsTmpl     = "extensions.go.tmpl"
 
 	// content.go generation
 	heuristicsRuby  = ".linguist/lib/linguist/heuristics.rb"
@@ -66,7 +68,7 @@ func main() {
 	}
 
 	argsList := []*generatorArgs{
-		&generatorArgs{languagesYAML, langFile, languagesTmplPath, languagesTmpl, commit, generator.Languages},
+		&generatorArgs{languagesYAML, extensionsFile, extensionsTmplPath, extensionsTmpl, commit, generator.Extensions},
 		&generatorArgs{heuristicsRuby, contentFile, contentTmplPath, contentTmpl, commit, generator.Heuristics},
 		&generatorArgs{vendorYAML, vendorFile, vendorTmplPath, vendorTmpl, commit, generator.Vendor},
 		&generatorArgs{documentationYAML, documentationFile, documentationTmplPath, documentationTmpl, commit, generator.Documentation},
